Remove associated features
These were somehow getting through osmfilter despite them not having the required tags. Probably associated with a polygon that does.

+++ update.pl
@@ -51,11 +51,11 @@ if(!-e $latest || $tstamp lt $yesterday){
 
 	if(-e $latest){
 		# Move the file
-		print "Move $latest > $old\n";
+		msg("Move <cyan>$latest<none> > <cyan>$old<none>\n");
 		`mv $latest $old`;
 	}
 
-	print "Download file from $url to $latest\n";
+	msg("Download file from <green>$url<none> to <cyan>$latest<none>\n");
 	`wget -q --no-check-certificate -O $latest "$url"`;
 }
 
@@ -63,7 +63,7 @@ if(-e $latest){
 
 	$tstamp = `osmconvert $latest --out-timestamp`;
 	$tstamp =~ s/[\n\r]//g;
-	print "PBF file last updated: $tstamp\n";
+	msg("PBF file last updated: <yellow>$tstamp<none>\n");
 
 #	# Update file
 #	print "Updating...\n";
@@ -74,25 +74,25 @@ if(-e $latest){
 #	}
 
 	if(-e $old){
-		print "Remove old version from $old\n";
+		msg("Remove old version from <cyan>$old<none>\n");
 		`rm $old`;
 	}
 
 	foreach $a (keys(%{$config->{'areas'}})){
 
-		print "$a:\n";
+		msg("Area <yellow>$a<none>:\n");
 				
 		if(!-e $basedir.$config->{'areas'}{$a}{'poly'}){
 			error("No polygon file $config->{'areas'}{$a}{'poly'}\n");
 		}
 		$arealatest = $tdir.$a."-latest.o5m";
 		# Make area extract
-		print "osmconvert $latest -B=$basedir$config->{'areas'}{$a}{'poly'} -o=$arealatest.\n";
+		msg("\tosmconvert $latest -B=$basedir$config->{'areas'}{$a}{'poly'} -o=$arealatest.\n");
 		`osmconvert $latest -B=$basedir$config->{'areas'}{$a}{'poly'} -o=$arealatest`;
 		
 
-		foreach $l (keys(%{$config->{'areas'}{$a}{'layers'}})){
-			print "layer $l\n";
+		foreach $l (sort(keys(%{$config->{'areas'}{$a}{'layers'}}))){
+			msg("\tLayer <yellow>$l<none>:\n");
 
 			# Extract layer
 			$layer = $tdir.$a."-$l.osm";
@@ -101,8 +101,8 @@ if(-e $latest){
 			if($config->{'areas'}{$a}{'layers'}{$l}{'keep'}){ $modify .= " --keep=\"$config->{'areas'}{$a}{'layers'}{$l}{'keep'}\""; }
 			if($config->{'areas'}{$a}{'layers'}{$l}{'drop'}){ $modify .= " --drop=\"$config->{'areas'}{$a}{'layers'}{$l}{'drop'}\""; }
 
-			print "osmfilter $arealatest $modify -o=$layer\n";
-			`osmfilter $arealatest $modify -o=$layer`;
+			msg("\t\tosmfilter $arealatest $modify -o=$layer\n");
+			`osmfilter $arealatest $modify -o=$layer 2>&1`;
 			$ldir = $basedir."layers/";
 			if(!-d $ldir){
 				`mkdir $ldir`;
@@ -115,18 +115,62 @@ if(-e $latest){
 
 			$geojson = $ldir."$l.geojson";
 
-			saveGeoJSONFeatures($layer,$geojson,$tstamp);
-			print "Remove $layer ";
-			`rm $layer`;
+			#if($layer =~ /craft/){
+			saveGeoJSONFeatures($layer,$geojson,$tstamp,$config->{'areas'}{$a}{'layers'}{$l});
+			#}
+			if(-e $layer){
+				msg("\t\tRemove $layer\n");
+				`rm $layer`;
+			}
 		}
-		`rm $arealatest`;
+		if(-e $arealatest){
+			`rm $arealatest`;
+		}
 	}
 }
 
-exit;
 
+sub checkFeature {
+	my $f = shift;
+	my $l = shift;
+	my @keep,$key,$value,$oldkey,$b,$i,$ok,$str;
 
-
+	$str = $l->{'keep'};
+	$str =~ s/ or / /g;
+	
+	my @bits = split(/\s/,$str);
+	for($b = 0; $b < @bits; $b++){
+		($key,$value) = split(/=/,$bits[$b]);
+		if(!$key){ $key = $oldkey; }
+		$value =~ s/ $//g;
+		push(@keep,{'key'=>$key,'value'=>$value});
+		$oldkey = $key;
+	}
+	$ok = 0;
+	for($i = 0; $i < @keep; $i++){
+		$value = $keep[$i]->{'value'};
+		$value =~ s/\*/\.\*/g;
+		if(defined($f->{'properties'}{$keep[$i]->{'key'}})){
+			if($value){
+				if($f->{'properties'}{$keep[$i]->{'key'}} =~ /$value/){
+					$ok++;
+				}
+			}else{
+				$ok++;
+			}
+		}
+		if(defined($f->{'properties'}{'other_tags'}{$keep[$i]->{'key'}})){
+			if($value){
+				if($f->{'properties'}{'other_tags'}{$keep[$i]->{'key'}} =~ /$value/){
+					$ok++;
+				}
+			}else{
+				$ok++;
+			}
+		}
+	}
+	return $ok;
+}
 
 
 
@@ -134,9 +178,10 @@ sub saveGeoJSONFeatures {
 
 	my (@types,@features,$t,@lines,$perl_scalar,$n,$i,$geojson,$txt,$bdir,$ini);
 	
-	my $osm = $_[0];
-	my $ofile = $_[1];
-	my $tstamp = $_[2];
+	my $osm = shift;
+	my $ofile = shift;
+	my $tstamp = shift;
+	my $props = shift;
 
 	$bdir = $osm;
 	$bdir =~ s/([^\/]+)$//;
@@ -149,23 +194,25 @@ sub saveGeoJSONFeatures {
 	foreach $t (@types){
 		$gfile = $bdir."temp-$t.geojson";
 		$ini = $basedir."osmconf.ini";
-		`ogr2ogr -overwrite --config OSM_CONFIG_FILE $ini -skipfailures -f GeoJSON $gfile $osm $t`;
+		`ogr2ogr -overwrite --config OSM_CONFIG_FILE $ini -skipfailures -f GeoJSON $gfile $osm $t 2>&1`;
 		open(FILE,$gfile);
 		@lines = <FILE>;
 		close(FILE);
 		$perl_scalar = $coder->decode(join("",@lines));
-#		print Dumper $perl_scalar;
 		$n = @{$perl_scalar->{'features'}};
 		for($i = 0; $i < $n; $i++){
 			# Process properties->other_tags into a structure
-			if($perl_scalar->{'features'}[$i]{'properties'}{'other_tags'}){
+			if(defined($perl_scalar->{'features'}[$i]{'properties'}{'other_tags'})){
 				$perl_scalar->{'features'}[$i]{'properties'}{'other_tags'} =~ s/\=\>/\:/g;
 				$perl_scalar->{'features'}[$i]{'properties'}{'other_tags'} = $coder->decode("{".$perl_scalar->{'features'}[$i]{'properties'}{'other_tags'}."}");
 			}
-			
-			push(@features,$perl_scalar->{'features'}[$i]);
+			if(checkFeature($perl_scalar->{'features'}[$i],$props)){
+				push(@features,$perl_scalar->{'features'}[$i]);
+			}
 		}
-		`rm $gfile`;
+		if(-e $gfile){
+			`rm $gfile`;
+		}
 	}
 	$n = @features;
 	$geojson = {'type'=>'FeatureCollection','features'=>\@features};
@@ -173,11 +220,12 @@ sub saveGeoJSONFeatures {
 	$txt = $coder->canonical(1)->encode($geojson);
 	$txt =~ s/(\{ ?"geometry":)/\n\t$1/gi;
 	$txt =~ s/(\],"type":"FeatureCollection")[^\}]*?\}/\n$1\}/;
+	$txt =~ s/,"other_tags":\{\}//g;
 
 	open(GEO,">",$ofile);
 	print GEO $txt;
 	close(GEO);	
-	print "$n features in $ofile\n";
+	msg("\t\t<yellow>$n<none> features in <cyan>$ofile<none>\n");
 
 	return;
 }
@@ -204,13 +252,41 @@ sub loadConf {
 }
 
 
-sub error {
+
+sub msg {
 	my $str = $_[0];
-	print "Content-type: text/html\n\n";
-	print "Error: $str\n";
-	exit;
+	my $dest = $_[1]||"STDOUT";
+	
+	my %colours = (
+		'black'=>"\033[0;30m",
+		'red'=>"\033[0;31m",
+		'green'=>"\033[0;32m",
+		'yellow'=>"\033[0;33m",
+		'blue'=>"\033[0;34m",
+		'magenta'=>"\033[0;35m",
+		'cyan'=>"\033[0;36m",
+		'white'=>"\033[0;37m",
+		'none'=>"\033[0m"
+	);
+	foreach my $c (keys(%colours)){ $str =~ s/\< ?$c ?\>/$colours{$c}/g; }
+	if($dest eq "STDERR"){
+		print STDERR $str;
+	}else{
+		print STDOUT $str;
+	}
 }
 
+sub error {
+	my $str = $_[0];
+	$str =~ s/(^[\t\s]*)/$1<red>ERROR:<none> /;
+	msg($str,"STDERR");
+}
+
+sub warning {
+	my $str = $_[0];
+	$str =~ s/(^[\t\s]*)/$1<yellow>WARNING:<none> /;
+	msg($str,"STDERR");
+}
 
 
 
@@ -220,7 +296,7 @@ sub tidyTemporaryFiles {
 	my $tstamp = substr($_[0],0,10);
 	my ($dh,$filename,@files,$i,@lines,$line);
 
-	print "Tidy $tstamp\n";
+	msg("Tidy <yellow>$tstamp<none>\n");
 	
 	opendir($dh,$tdir);
 	while( ($filename = readdir($dh))) {
@@ -231,7 +307,7 @@ sub tidyTemporaryFiles {
 			foreach $line (@lines){
 				if($line =~ /timestamp=([0-9]{4}-[0-9]{2}-[0-9]{2})/){
 					if($1 lt $tstamp){
-						print "Remove $filename\n";
+						msg("Remove <cyan>$filename<none>\n");
 #						`rm $tdir$filename`;
 					}
 				}
